Fix the click-at-outset
Some of our shapers didn't map 0->0. Each of those had
a DC blocker so that long playing silence didn't produce DC
but at onset, the DC blocker hadn't had time to catch up.
This change seeds the DC blocker so that at init, the 'prior' value
it observes is the input-to-the-blocker for 0, defacto making
no click on change at silence.